Basic Refactoring of CenterLines class to move extra private functions to a utility class so as to improve clutter and readability of the CenterLines public functions, as well as improving the maintainability of the class moving forward. Also updated tests to reflect changes in imports to ensure that code stays in working order.
